Pixy Frame Interpreter using Arduino
This file goes through every frame that the pixy cam dumps and puts the image through a hue saturation value filter that allows me to track an object at a certain value on the hsv spectrum. It then outputs xy and length and height values of the object to allow for tracking. This is over the data protocol of I2C.